feat: create plugin with db init command
